agregue una nueva linea de codigo

+++ main.js
@@ -1,23 +1,31 @@
-
+    
     const input = document.getElementById("input");
     const btn_todo = document.getElementById("btn_todo");
     const container_todo = document.querySelector(".container_todo");
 
     const guardarTareas = () => {
         const tarea = {
-            input_tarea: input.value
+            input_tarea: input.value.trim()
         }
-
+    
         if (localStorage.getItem("tareas") === null) {
             let arreglo = [];
             arreglo.push(tarea);
             localStorage.setItem("tareas", JSON.stringify(arreglo));
         } else {
             let obtener = JSON.parse(localStorage.getItem("tareas"));
-            obtener.push(tarea);
-            localStorage.setItem("tareas", JSON.stringify(obtener));
-        } 
-        
+    
+            // Verificar si el elemento ya existe
+            const existe = obtener.some(item => item.input_tarea.toLowerCase() === tarea.input_tarea.toLowerCase());
+    
+            if (existe) {
+                alert("El elemento ya existe en la lista.");
+            } else {
+                obtener.push(tarea);
+                localStorage.setItem("tareas", JSON.stringify(obtener));
+            }
+        }
+    
         mostrarTareas();
         input.value = "";
     }
@@ -63,4 +71,4 @@
         }
     });
 
-
+document.addEventListener("DOMContentLoaded", mostrarTareas);
